Fix bugs, optimize GhostLog, QtSimulator, and components support for lvgl.
1. Add lv_lib_png.
2. Refactor GhostLog and fix bugs.
3. Support dynamic adjustment of QtSimulator resolution.
4. Optimize GhostClock.
5. Support lvgl file system.
6. Support output logs to UI of QtSimulator.

+++ Ghost/Drivers/Clock/GhostClock.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "GhostPlatformErrorDefine.h"
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -34,8 +35,8 @@ extern "C" {
 	/// Get the time from startup to now.
 	/// </summary>
 	/// <returns></returns>
-	GhostError_t GhostGetCurrentMilliseconds(int* Milliseconds);
-	GhostError_t GhostGetCurrentMicroseconds(int* Microseconds);
+	GhostError_t GhostGetCurrentMilliseconds(double* Milliseconds);
+	GhostError_t GhostGetCurrentMicroseconds(double* Microseconds);
 
 	/// <summary>
 	/// Sleep functions.
